Move AcceptanceHelper to the  codeception default location for helpers (#111)
* Use default path for the AcceptanceHelper

* Use codeception defaults for helper files

* Order enabled modules alphabetical
