+++ tests/codeception/acceptance.suite.dist.yml
@@ -7,10 +7,10 @@
 class_name: AcceptanceTester
 modules:
     enabled:
-        - JoomlaBrowser
-        - AcceptanceHelper
         - Asserts
-        - \Helper\JoomlaDb
+        - JoomlaBrowser
+        - Helper\Acceptance
+        - Helper\JoomlaDb
     config:
         JoomlaBrowser:
             url: 'http://localhost/tests/codeception/joomla-cms3'     # the url that points to the joomla installation at /tests/system/joomla-cms
@@ -30,10 +30,10 @@ modules:
             sample data: 'Default English (GB) Sample Data'    # Default Sample Data
             admin email: 'admin@mydomain.com'      # email Id of the Admin
             language: 'English (United Kingdom)'   # Language in which you want the Application to be Installed
-        AcceptanceHelper:
+        Helper\Acceptance:
             url: 'http://localhost/tests/codeception/joomla-cms3' # the url that points to the joomla installation at /tests/system/joomla-cms - we need it twice here
             MicrosoftEdgeInsiders: false                  # set this to true, if you are on Windows Insiders
-        \Helper\JoomlaDb:                                 # http://codeception.com/docs/modules/Db
+        Helper\JoomlaDb:                                 # http://codeception.com/docs/modules/Db
           dsn: 'mysql:host=localhost;dbname=test_joomla'  # PDO DSN
           user: 'root'                                    # MySQL Server user
           password: ''                                    # MySQL Server password
